+++ Source/OxyPlot/Axes/Axis.cs
@@ -77,6 +77,10 @@ namespace OxyPlot.Axes
             this.MaximumPadding = 0.01;
             this.MinimumRange = 0;
             this.MaximumRange = double.PositiveInfinity;
+            this.MinimumDataMargin = 0;
+            this.MaximumDataMargin = 0;
+            this.MinimumMargin = 0;
+            this.MaximumMargin = 0;
 
             this.TickStyle = TickStyle.Outside;
             this.TicklineColor = OxyColors.Black;
@@ -173,6 +177,22 @@ namespace OxyPlot.Axes
         /// Otherwise, if <see cref="Minimum" /> is not <c>NaN</c>, this value will be defined by <see cref="Minimum" />.
         /// Otherwise this value will be defined by the minimum (+padding) of the data.</remarks>
         public double ActualMinimum { get; protected set; }
+
+        /// <summary>
+        /// Gets or sets the maximum displayed value on the axis, as determined by the <see cref="ActualMaximum"/> and <see cref="MaximumDataMargin"/>.
+        /// </summary>
+        /// <remarks>
+        /// The value is refreshed by <see cref="UpdateTransform(OxyRect)"/>, which is called before any plot elements are rendered.
+        /// </remarks>
+        public double ClipMaximum { get; protected set; }
+
+        /// <summary>
+        /// Gets or sets the minimum displayed value on the axis, as determined by the <see cref="ActualMinimum"/> and <see cref="MinimumDataMargin"/>.
+        /// </summary>
+        /// <remarks>
+        /// The value is refreshed by <see cref="UpdateTransform(OxyRect)"/>, which is called before any plot elements are rendered.
+        /// </remarks>
+        public double ClipMinimum { get; protected set; }
 
         /// <summary>
         /// Gets or sets the actual minor step.
@@ -382,6 +402,18 @@ namespace OxyPlot.Axes
         public double MaximumPadding { get; set; }
 
         /// <summary>
+        /// Gets or sets the screen-space data margin at the maximum. The default value is <c>0</c>.
+        /// </summary>
+        /// <value>The number of device independent units to included between the <see cref="ClipMaximum"/> and <see cref="ActualMaximum"/>.</value>
+        public double MaximumDataMargin { get; set; }
+
+        /// <summary>
+        /// Gets or sets the screen-space margin at the maximum. The default value is <c>0</c>.
+        /// </summary>
+        /// <value>The number of device independent units to be left empty between the axis and the <see cref="EndPosition"/>.</value>
+        public double MaximumMargin { get; set; }
+
+        /// <summary>
         /// Gets or sets the maximum range of the axis. Setting this property ensures that <c>ActualMaximum-ActualMinimum &lt; MaximumRange</c>. The default value is <c>double.PositiveInfinity</c>.
         /// </summary>
         public double MaximumRange { get; set; }
@@ -406,6 +438,18 @@ namespace OxyPlot.Axes
         /// </summary>
         /// <remarks>A value of 0.01 gives 1% more space on the minimum end of the axis. This property is not used if the <see cref="Minimum" /> property is set.</remarks>
         public double MinimumPadding { get; set; }
+
+        /// <summary>
+        /// Gets or sets the screen-space data margin at the minimum. The default value is <c>0</c>.
+        /// </summary>
+        /// <value>The number of device independent units to included between the <see cref="ClipMinimum"/> and <see cref="ActualMinimum"/>.</value>
+        public double MinimumDataMargin { get; set; }
+
+        /// <summary>
+        /// Gets or sets the screen-space margin at the minimum. The default value is <c>0</c>.
+        /// </summary>
+        /// <value>The number of device independent units to be left empty between the axis the <see cref="StartPosition"/>.</value>
+        public double MinimumMargin { get; set; }
 
         /// <summary>
         /// Gets or sets the minimum range of the axis. Setting this property ensures that <c>ActualMaximum-ActualMinimum > MinimumRange</c>. The default value is <c>0</c>.
@@ -557,7 +601,7 @@ namespace OxyPlot.Axes
         public double TitleFontWeight { get; set; }
 
         /// <summary>
-        /// Gets or sets the format string used for formatting the title and unit when <see cref="Unit" /> is defined. 
+        /// Gets or sets the format string used for formatting the title and unit when <see cref="Unit" /> is defined.
         /// The default value is "{0} [{1}]", where {0} refers to the <see cref="Title" /> and {1} refers to the <see cref="Unit" />.
         /// </summary>
         /// <remarks>If <see cref="Unit" /> is <c>null</c>, the actual title is defined by <see cref="Title" /> only.</remarks>
@@ -662,7 +706,7 @@ namespace OxyPlot.Axes
         protected double ViewMinimum { get; set; }
 
         /// <summary>
-        /// Converts the value of the specified object to a double precision floating point number. DateTime objects are converted using DateTimeAxis.ToDouble and TimeSpan objects are converted using TimeSpanAxis.ToDouble
+        /// Converts the value of the specified object to a double precision floating point number. DateTime objects are converted using DateTimeAxis.ToDouble and TimeSpan objects are converted using TimeSpanAxis.ToDouble.
         /// </summary>
         /// <param name="value">The value.</param>
         /// <returns>The floating point number value.</returns>
@@ -717,8 +761,8 @@ namespace OxyPlot.Axes
         public virtual void GetTickValues(
             out IList<double> majorLabelValues, out IList<double> majorTickValues, out IList<double> minorTickValues)
         {
-            minorTickValues = this.CreateTickValues(this.ActualMinimum, this.ActualMaximum, this.ActualMinorStep);
-            majorTickValues = this.CreateTickValues(this.ActualMinimum, this.ActualMaximum, this.ActualMajorStep);
+            minorTickValues = this.CreateTickValues(this.ClipMinimum, this.ClipMaximum, this.ActualMinorStep);
+            majorTickValues = this.CreateTickValues(this.ClipMinimum, this.ClipMaximum, this.ActualMajorStep);
             majorLabelValues = majorTickValues;
 
             minorTickValues = AxisUtilities.FilterRedundantMinorTicks(majorTickValues, minorTickValues);
@@ -841,6 +885,9 @@ namespace OxyPlot.Axes
             var marginRight = 0d;
             var marginBottom = 0d;
 
+            var minOuterMargin = Math.Max(0, this.IsReversed ? this.MaximumMargin : this.MinimumMargin);
+            var maxOuterMargin = Math.Max(0, this.IsReversed ? this.MinimumMargin : this.MaximumMargin);
+
             var margin = this.TickStyle switch
             {
                 TickStyle.Outside => this.MajorTickSize,
@@ -888,12 +935,12 @@ namespace OxyPlot.Axes
                     case AxisPosition.Right:
                         if (reachesMinPosition)
                         {
-                            marginBottom = maximumTextSize.Height / 2;
+                            marginBottom = Math.Max(0, (maximumTextSize.Height / 2) - minOuterMargin);
                         }
 
                         if (reachesMaxPosition)
                         {
-                            marginTop = maximumTextSize.Height / 2;
+                            marginTop = Math.Max(0, (maximumTextSize.Height / 2) - maxOuterMargin);
                         }
 
                         break;
@@ -901,12 +948,12 @@ namespace OxyPlot.Axes
                     case AxisPosition.Bottom:
                         if (reachesMinPosition)
                         {
-                            marginLeft = maximumTextSize.Width / 2;
+                            marginLeft = Math.Max(0, (maximumTextSize.Width / 2) - minOuterMargin);
                         }
 
                         if (reachesMaxPosition)
                         {
-                            marginRight = maximumTextSize.Width / 2;
+                            marginRight = Math.Max(0, (maximumTextSize.Width / 2) - maxOuterMargin);
                         }
 
                         break;
@@ -939,16 +986,16 @@ namespace OxyPlot.Axes
                         var screenMinY = Math.Min(this.ScreenMin.Y, this.ScreenMax.Y);
                         var screenMaxY = Math.Max(this.ScreenMin.Y, this.ScreenMax.Y);
 
-                        marginTop = Math.Max(0, screenMinY - minLabelPosition + (minLabelSize.Height / 2));
-                        marginBottom = Math.Max(0, maxLabelPosition - screenMaxY + (maxLabelSize.Height / 2));
+                        marginTop = Math.Max(0, screenMinY - minLabelPosition + (minLabelSize.Height / 2) - minOuterMargin);
+                        marginBottom = Math.Max(0, maxLabelPosition - screenMaxY + (maxLabelSize.Height / 2) - maxOuterMargin);
                         break;
                     case AxisPosition.Top:
                     case AxisPosition.Bottom:
                         var screenMinX = Math.Min(this.ScreenMin.X, this.ScreenMax.X);
                         var screenMaxX = Math.Max(this.ScreenMin.X, this.ScreenMax.X);
 
-                        marginLeft = Math.Max(0, screenMinX - minLabelPosition + (minLabelSize.Width / 2));
-                        marginRight = Math.Max(0, maxLabelPosition - screenMaxX + (maxLabelSize.Width / 2));
+                        marginLeft = Math.Max(0, screenMinX - minLabelPosition + (minLabelSize.Width / 2) - minOuterMargin);
+                        marginRight = Math.Max(0, maxLabelPosition - screenMaxX + (maxLabelSize.Width / 2) - maxOuterMargin);
                         break;
                 }
             }
@@ -1059,8 +1106,8 @@ namespace OxyPlot.Axes
                 "{0}({1}, {2}, {3}, {4})",
                 this.GetType().Name,
                 this.Position,
-                this.ActualMinimum,
-                this.ActualMaximum,
+                this.ClipMinimum,
+                this.ClipMaximum,
                 this.ActualMajorStep);
         }
 
@@ -1154,13 +1201,14 @@ namespace OxyPlot.Axes
                 }
             }
 
-            this.ViewMaximum = newMaximum;
-            this.ViewMinimum = newMinimum;
+            this.ClipMaximum = this.ViewMaximum = newMaximum;
+            this.ClipMinimum = this.ViewMinimum = newMinimum;
             this.UpdateActualMaxMin();
 
             var deltaMinimum = this.ActualMinimum - oldMinimum;
             var deltaMaximum = this.ActualMaximum - oldMaximum;
 
+            this.ActualMaximumAndMinimumChangedOverride();
             this.OnAxisChanged(new AxisChangedEventArgs(AxisChangeTypes.Zoom, deltaMinimum, deltaMaximum));
         }
 
@@ -1247,7 +1295,7 @@ namespace OxyPlot.Axes
         /// <param name="factor">The zoom factor.</param>
         public virtual void ZoomAtCenter(double factor)
         {
-            double sx = (this.Transform(this.ActualMaximum) + this.Transform(this.ActualMinimum)) * 0.5;
+            double sx = (this.Transform(this.ClipMaximum) + this.Transform(this.ClipMinimum)) * 0.5;
             var x = this.InverseTransform(sx);
             this.ZoomAt(factor, x);
         }
@@ -1360,17 +1408,37 @@ namespace OxyPlot.Axes
             double y0 = bounds.Bottom;
             double y1 = bounds.Top;
 
-            this.ScreenMin = new ScreenPoint(x0, y1);
-            this.ScreenMax = new ScreenPoint(x1, y0);
-
             double a0 = this.IsHorizontal() ? x0 : y0;
             double a1 = this.IsHorizontal() ? x1 : y1;
 
             double dx = a1 - a0;
             a1 = a0 + (this.EndPosition * dx);
             a0 = a0 + (this.StartPosition * dx);
+
+            double marginSign = (this.IsHorizontal() ^ this.IsReversed) ? 1.0 : -1.0;
+
+            if (this.MinimumMargin > 0)
+            {
+                a0 += this.MinimumMargin * marginSign;
+            }
+
+            if (this.MaximumMargin > 0)
+            {
+                a1 -= this.MaximumMargin * marginSign;
+            }
+
             this.ScreenMin = new ScreenPoint(a0, a1);
             this.ScreenMax = new ScreenPoint(a1, a0);
+
+            if (this.MinimumDataMargin > 0)
+            {
+                a0 += this.MinimumDataMargin * marginSign;
+            }
+
+            if (this.MaximumDataMargin > 0)
+            {
+                a1 -= this.MaximumDataMargin * marginSign;
+            }
 
             if (this.ActualMaximum - this.ActualMinimum < double.Epsilon)
             {
@@ -1402,6 +1470,33 @@ namespace OxyPlot.Axes
             }
 
             this.SetTransform(newScale, newOffset);
+
+            if (this.MinimumDataMargin > 0)
+            {
+                this.ClipMinimum = this.InverseTransform(a0 - (this.MinimumDataMargin * marginSign));
+            }
+            else
+            {
+                this.ClipMinimum = this.ActualMinimum;
+            }
+
+            if (this.MaximumDataMargin > 0)
+            {
+                this.ClipMaximum = this.InverseTransform(a1 + (this.MaximumDataMargin * marginSign));
+            }
+            else
+            {
+                this.ClipMaximum = this.ActualMaximum;
+            }
+
+            this.ActualMaximumAndMinimumChangedOverride();
+        }
+
+        /// <summary>
+        /// Invoked when <see cref="ActualMinimum"/>, <see cref="ActualMaximum"/>, <see cref="ClipMinimum"/>, and <see cref="ClipMaximum"/> are changed.
+        /// </summary>
+        protected virtual void ActualMaximumAndMinimumChangedOverride()
+        {
         }
 
         /// <summary>
@@ -1456,7 +1551,7 @@ namespace OxyPlot.Axes
         /// <param name="step">The interval.</param>
         /// <param name="maxTicks">The maximum number of ticks (optional). The default value is 1000.</param>
         /// <returns>A sequence of values.</returns>
-        /// <exception cref="System.ArgumentException">Step cannot be zero or negative.;step</exception>
+        /// <exception cref="System.ArgumentException">Step cannot be zero or negative.</exception>
         protected virtual IList<double> CreateTickValues(double from, double to, double step, int maxTicks = 1000)
         {
             return AxisUtilities.CreateTickValues(from, to, step, maxTicks);